make mutation logic standalone, ensure clusters created with kubeapi get auto-enabled external CCM
